requirements.txt
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
streamlit>=1.20.0
requests>=2.28.0
numpy>=1.23.0
Pillow>=9.3.0
opencv-python>=4.6.0
tensorflow-cpu>=2.10.0
python-multipart>=0.0.5
protobuf<4.0.0,>=3.20.0
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Streamlit and Visualization
streamlit==1.28.2
plotly==5.18.0

# HTTP Requests
requests==2.31.0

# Data Processing
numpy==1.24.3
pandas==2.1.3

# Optional but recommended
python-dotenv==1.0.0
